Update Global Install Instructions
Using `npm i -g` is generally a bad idea in many secure environments as
it requires root or sudo access to install a global module. Installing
the module globally also prevents ongoing development (without frequent
reinstalls).

Instead, using `npm link` will create a symlink for the project in the
_user's_ path. It provides the same general effect (the user can run
`create-10up` from the command line and use the tool), but without
requiring special permissions.

As the `link` is a symlink, any changes to the module on disk are
automatically reflected by the installed tool as well.